Make Cypher return 64-bit integers uniformly

+++ community/cypher/cypher-compiler-3.0/src/main/scala/org/neo4j/cypher/internal/compiler/v3_0/commands/expressions/GetDegree.scala
@@ -30,7 +30,7 @@ import org.neo4j.graphdb.Node
 
 case class GetDegree(node: Expression, typ: Option[KeyToken], direction: SemanticDirection) extends NullInNullOutExpression(node) {
 
-  val getDegree: (QueryContext, Long) => Int = typ match {
+  val getDegree: (QueryContext, Long) => Long = typ match {
     case None    => (qtx, node) => qtx.nodeGetDegree(node, direction)
     case Some(t) => (qtx, node) => t.getOptId(qtx) match {
       case None            => 0
